server/nlp_sql/nlp_to_sql/requirements.txt
flask
flask_cors
pandas
llama-cpp-python
langchain
